requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.6
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
django-admin-honeypot==1.1.0
django-appconf==1.0.3
django-autocomplete-light==3.3.2
django-cors-headers==2.5.0
django-decorator-include==2.1
django-djconfig==0.8.0
django-haystack==2.8.1
django-infinite-scroll-pagination==0.2.0
django-oauth-toolkit==1.2.0
django-phonenumber-field==2.2.0
django-rest-knox==3.6.0
django-select2==6.3.1
django-url-filter==0.3.12
django-vote==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
enum-compat==0.0.2
httplib2==0.12.1
idna==2.8
mistune==0.8.4
oauthlib==3.0.1
olefile==0.46
phonenumbers==8.10.7
pilkit==2.0
Pillow==5.4.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
uni-slugify==0.1.4
Unidecode==1.0.23
urllib3==1.24.1
Whoosh==2.7.4
